+++ ui/src/organizations/components/TaskRow.tsx
@@ -7,6 +7,9 @@ import {
   IndexList,
   ConfirmationButton,
   Alignment,
+  Button,
+  IconFont,
+  ComponentColor,
 } from 'src/clockface'
 
 // Api
@@ -17,6 +20,7 @@ interface Props {
   task: Task
   onDelete: (taskID: string) => void
   onUpdate: (task: Task) => void
+  onClone: (task: Task) => void
 }
 
 export default class TaskRow extends PureComponent<Props> {
@@ -35,6 +39,13 @@ export default class TaskRow extends PureComponent<Props> {
           </IndexList.Cell>
           <IndexList.Cell>{task.name}</IndexList.Cell>
           <IndexList.Cell revealOnHover={true} alignment={Alignment.Right}>
+            <Button
+              size={ComponentSize.ExtraSmall}
+              color={ComponentColor.Secondary}
+              text="Clone"
+              icon={IconFont.Duplicate}
+              onClick={this.handleClone}
+            />
             <ConfirmationButton
               size={ComponentSize.ExtraSmall}
               text="Delete"
@@ -50,6 +61,10 @@ export default class TaskRow extends PureComponent<Props> {
   private handleUpdateTask = (name: string) => {
     const {onUpdate, task} = this.props
     onUpdate({...task, name})
+  }
+
+  private handleClone = (): void => {
+    this.props.onClone(this.props.task)
   }
 
   private handleDeleteTask = (): void => {