+++ ui/src/utils/api.ts
@@ -1,11 +1,4 @@
-import {
-  DashboardsApi,
-  CellsApi,
-  ViewsApi,
-  SourcesApi,
-  QueryApi,
-  ProtosApi,
-} from 'src/api'
+import {DashboardsApi, QueryApi} from 'src/api'
 
 import Client from '@influxdata/influx'
 
@@ -13,9 +6,5 @@ const basePath = '/api/v2'
 
 export const client = new Client(basePath)
 
-export const viewsAPI = new ViewsApi({basePath})
 export const dashboardsAPI = new DashboardsApi({basePath})
-export const cellsAPI = new CellsApi({basePath})
-export const sourcesAPI = new SourcesApi({basePath})
 export const queryAPI = new QueryApi({basePath})
-export const protosAPI = new ProtosApi({basePath})