+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## UI Improvements
 1. [11683](https://github.com/influxdata/influxdb/pull/11683): Change the wording for the plugin config form button to Done
+1. [11689](https://github.com/influxdata/influxdb/pull/11689): Change the wording for the Collectors configure step button to Create and Verify
 
 ## v2.0.0-alpha.1 [2019-01-23]