+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v2.0.0-alpha.2 [unreleased]
+
+## Features
+1. [11693](https://github.com/influxdata/influxdb/pull/11693): Save the $INFLUX_TOKEN environmental variable in telegraf configs
+
+## Bug Fixes
+1. [11678](https://github.com/influxdata/influxdb/pull/11678): Update the System Telegraf Plugin bundle to include the swap plugin
+
+## UI Improvements
+1. [11683](https://github.com/influxdata/influxdb/pull/11683): Change the wording for the plugin config form button to Done
+1. [11689](https://github.com/influxdata/influxdb/pull/11689): Change the wording for the Collectors configure step button to Create and Verify
+
 ## v2.0.0-alpha.1 [2019-01-23]
 
 ### Release Notes